// WMS-UI/src/app/core/models/employee.ts
export interface Employee {
  employeeId: number;
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  gender: string;
  dob: string;
  doj: string;
  departmentId: number;
  departmentName: string;
  roleId: number;
  roleName: string;
  status: string;
  createdOn: string;
}